Move responsibility for running the tasks of the root build into `BuildTreeWorkGraph`.
Now the scheduled tasks of all builds in the build tree are managed and lifecycled in the same way.
