--- subprojects/composite-builds/src/main/java/org/gradle/composite/internal/RootBuildController.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2021 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.composite.internal;
-
-import org.gradle.internal.build.BuildState;
-
-import java.util.concurrent.ExecutorService;
-import java.util.function.Consumer;
-
-class RootBuildController extends AbstractBuildController {
-    public RootBuildController(BuildState build) {
-        super(build);
-    }
-
-    @Override
-    protected void doStartTaskExecution(ExecutorService executorService) {
-        // This is started via another path
-    }
-
-    @Override
-    protected void doAwaitTaskCompletion(Consumer<? super Throwable> taskFailures) {
-    }
-}
