+++ subprojects/internal-testing/src/main/groovy/org/gradle/test/fixtures/work/TestWorkerLeaseService.groovy
@@ -55,7 +55,7 @@ class TestWorkerLeaseService implements WorkerLeaseService {
 
     @Override
     void withSharedLease(WorkerLeaseRegistry.WorkerLease sharedLease, Runnable action) {
-        throw new UnsupportedOperationException("not implemented yet")
+        action.run()
     }
 
     @Override
